pyproject.toml
[tool.poetry]
name = "drf-cars"
version = "0.1.0"
description = ""
authors = ["tetianaklius <tetyanaklyus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.1.0"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
django-filter = "^24.3"
better-profanity = "^0.7.0"
django-guardian = "^2.4.0"
djangochannelsrestframework = "^1.2.0"
daphne = "^4.1.2"
channels-redis = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
